requirements.txt
streamlit
streamlit-option-menu
google-generativeai
python-dotenv
logger
PyPDF2
langchain_google_genai
textwrap3
youtube_transcript_api
langchain
langchain_community
streamlit_lottie
openpyxl
pygwalker
pandasai
pandas
-e .